Make id provider factory outer class
Also rename configurator interface to factory and remove unused method
