+++ src/us/kbase/auth2/lib/identity/IdentityProvider.java
@@ -8,8 +8,7 @@ import us.kbase.auth2.lib.exceptions.IdentityRetrievalException;
 /** A provider of OAuth2 identities, for example Google, Globus, Facebook etc.
  * @author gaprice@lbl.gov
  * 
- * @see IdentityProviderConfigurator
- * @see IdentityProviderSet
+ * @see IdentityProviderFactory
  *
  */
 public interface IdentityProvider {
