+++ src/us/kbase/test/auth2/kbase/GoogleIdentityProviderTest.java
@@ -27,8 +27,8 @@ import org.mockserver.model.ParameterBody;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import us.kbase.auth2.kbase.GoogleIdentityProvider;
-import us.kbase.auth2.kbase.GoogleIdentityProvider.GoogleIdentityProviderConfigurator;
+import us.kbase.auth2.kbase.GoogleIdentityProviderFactory;
+import us.kbase.auth2.kbase.GoogleIdentityProviderFactory.GoogleIdentityProvider;
 import us.kbase.auth2.lib.exceptions.IdentityRetrievalException;
 import us.kbase.auth2.lib.identity.IdentityProvider;
 import us.kbase.auth2.lib.identity.IdentityProviderConfig;
@@ -84,7 +84,7 @@ public class GoogleIdentityProviderTest {
 	static {
 		try {
 			CFG = new IdentityProviderConfig(
-					GoogleIdentityProviderConfigurator.class.getName(),
+					GoogleIdentityProviderFactory.class.getName(),
 					new URL("https://glogin.com"),
 					new URL("https://gsetapiurl.com"),
 					"gfoo",
@@ -98,8 +98,7 @@ public class GoogleIdentityProviderTest {
 	
 	@Test
 	public void simpleOperationsWithConfigurator() throws Exception {
-		final GoogleIdentityProviderConfigurator gc = new GoogleIdentityProviderConfigurator();
-		assertThat("incorrect provider name", gc.getProviderName(), is("Google"));
+		final GoogleIdentityProviderFactory gc = new GoogleIdentityProviderFactory();
 		
 		final IdentityProvider gip = gc.configure(CFG);
 		assertThat("incorrect provider name", gip.getProviderName(), is("Google"));
@@ -183,7 +182,7 @@ public class GoogleIdentityProviderTest {
 			throws IdentityProviderConfigurationException, MalformedURLException,
 			URISyntaxException {
 		return new IdentityProviderConfig(
-				GoogleIdentityProviderConfigurator.class.getName(),
+				GoogleIdentityProviderFactory.class.getName(),
 				new URL("https://glogin.com"),
 				new URL("http://localhost:" + mockClientAndServer.getPort()),
 				"gfoo",
@@ -378,7 +377,7 @@ public class GoogleIdentityProviderTest {
 	public void getIdentityWithLinkURL() throws Exception {
 		final String authCode = "authcode2";
 		final IdentityProviderConfig idconfig = new IdentityProviderConfig(
-				GoogleIdentityProviderConfigurator.class.getName(),
+				GoogleIdentityProviderFactory.class.getName(),
 				new URL("https://glogin2.com"),
 				new URL("http://localhost:" + mockClientAndServer.getPort()),
 				"someclient",
